[FIX] - 공통된 Test Func 통합 Class TestApis 수정
