+++ src/test/kotlin/sinhee/kang/tutorial/TestApis.kt
@@ -1,0 +1,72 @@
+package sinhee.kang.tutorial
+
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.databind.PropertyNamingStrategy
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.http.MediaType
+import org.springframework.test.context.ActiveProfiles
+import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers
+
+import sinhee.kang.tutorial.domain.auth.dto.request.SignInRequest
+import sinhee.kang.tutorial.domain.auth.dto.response.TokenResponse
+import sinhee.kang.tutorial.infra.redis.EmbeddedRedisConfig
+
+@SpringBootTest(classes = [TutorialApplication::class, EmbeddedRedisConfig::class],
+    webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@ActiveProfiles("test", "local")
+class TestApis(): DiVariables() {
+
+    protected fun requestBody(method: MockHttpServletRequestBuilder, obj: Any): String {
+        return mvc.perform(
+            method
+                .content(ObjectMapper()
+                    .setPropertyNamingStrategy(PropertyNamingStrategy.SNAKE_CASE)
+                    .writeValueAsString(obj))
+                .contentType(MediaType.APPLICATION_JSON_VALUE))
+            .andExpect(MockMvcResultMatchers.status().isOk)
+            .andReturn().response.contentAsString
+    }
+
+    protected fun requestBody(method: MockHttpServletRequestBuilder, obj: Any? = null, token: String?): String {
+        return mvc.perform(
+            method
+                .header("Authorization", token)
+                .content(ObjectMapper()
+                    .setPropertyNamingStrategy(PropertyNamingStrategy.SNAKE_CASE)
+                    .writeValueAsString(obj))
+                .contentType(MediaType.APPLICATION_JSON_VALUE))
+            .andExpect(MockMvcResultMatchers.status().isOk)
+            .andReturn().response.contentAsString
+    }
+
+    protected fun getAccessToken(signInRequest: SignInRequest): String {
+        val content = requestBody(post("/auth"), signInRequest)
+        val response = mappingResponse(content, TokenResponse::class.java) as TokenResponse
+
+        return "Bearer " + response.accessToken
+    }
+
+    protected fun generatePost(
+        method: MockHttpServletRequestBuilder = post("/posts"),
+        title: String = "title",
+        content: String = "content",
+        tags: String = "#tag",
+        token: String?
+    ): Int =
+        mvc.perform(
+            method
+                .header("Authorization", token)
+                .param("title", title)
+                .param("content", content)
+                .param("tags", tags))
+            .andExpect(MockMvcResultMatchers.status().isOk)
+            .andReturn().response.contentAsString.toInt()
+
+    protected fun mappingResponse(obj: String, cls: Class<*>): Any {
+        return objectMapper
+            .setPropertyNamingStrategy(PropertyNamingStrategy.SNAKE_CASE)
+            .readValue(obj, cls)
+    }
+}
